Move declare to index file, fix eslint config, eslint typescript
