--- .eslintrc
@@ -1,24 +1,0 @@
-{
-    "parser": "babel-eslint",
-    "env": {
-        "browser": true,
-        "node": true,
-        "es6": true
-    },
-    "parserOptions": {
-        "ecmaFeatures": {
-            "blockBindings": true,
-            "forOf": true
-        }
-    },
-    "rules": {
-        "no-unused-vars": ["error", { "vars": "all", "args": "after-used" }],
-        "comma-dangle": 0,
-        "indent": ["error", 4, {SwitchCase: 1}],
-        "quotes": [2, "double", "avoid-escape"],
-        "semi": [2, "always"],
-        "camelcase": [0],
-        "no-console": [0],
-        "object-curly-spacing": [0]
-    }
-}
