+++ features/steps/storage/google-drive-node.ts
@@ -4,14 +4,15 @@ import GoogleDriveAdapterFactory from "../../../lib/storage/src/GoogleDriveAdapt
 import sinon from 'sinon';
 import fs from "fs";
 import os from "os";
-import { GoogleDriveAdapter } from "../../../lib";
 import { expect } from "chai";
+import { Credentials } from "aws-sdk";
+import { GoogleDriveAdapter } from "../../../lib";
 
 After(async function (this: StorageWorld) {
     if (fs.existsSync(this.tokenFile)) {
         fs.unlinkSync(this.tokenFile)
     }
-});
+})
 
 Given('I have a credentials in json format', async function (this: StorageWorld) {
     this.tokenFile = `${os.tmpdir()}/google_token.json`;
@@ -29,18 +30,18 @@ Given('I have a credentials in json format', async function (this: StorageWorld)
             ]
         }
     }
-});
+})
 
 Given('google drive storage factory', async function (this: StorageWorld) {
     this.googleDriveAdapterFactory = new GoogleDriveAdapterFactory();
-    const fakeTokens = {
+    this.fakeToken = {
         "access_token": "fake_access_token",
         "refresh_token": "fake_refresh_token",
         "scope": "https://www.googleapis.com/auth/drive.file",
         "token_type": "Bearer",
-        "expiry_date": 1644338594649
-    };
-    sinon.replace(this.googleDriveAdapterFactory, 'getAccessToken' as any, sinon.fake.returns(fakeTokens));
+        "expiry_date": 1644338594649,
+    }
+    sinon.replace(this.googleDriveAdapterFactory, 'getAccessToken' as any, sinon.fake.returns(this.fakeToken));
 });
 
 When('I create storage', async function (this: StorageWorld) {
@@ -52,5 +53,9 @@ When('I create storage', async function (this: StorageWorld) {
 
 Then('I should get storage with google drive adapter', function (this: StorageWorld) {
     expect(fs.existsSync(this.tokenFile)).to.be.true
+
+    const token: Credentials = JSON.parse(fs.readFileSync(this.tokenFile).toString())
+    expect(token).to.deep.equal(this.fakeToken)
+
     expect(this.adapter).to.be.instanceof(GoogleDriveAdapter);
 });
