+++ lib/index.d.ts
@@ -1,4 +1,4 @@
-declare class Adapter {
+export declare class Adapter {
     open(): Promise<void>;
     close(): Promise<void>;
     /**
@@ -27,4 +27,32 @@ declare class Adapter {
     remove(key: string, options?: Object): Promise<boolean>;
 }
 
-export default Adapter
+export declare class GoogleDriveAdapter implements Adapter {
+    constructor(options?: any);
+    folder: any;
+    drive: any;
+
+    close(): Promise<void>;
+
+    open(): Promise<void>;
+
+    /**
+     * Store the passed value
+     *
+     * @param {string|Buffer} val
+     * @param {Object} options
+     * @returns {Promise<String>}
+     */
+    put(val: Uint8Array, options?: Object): Promise<string>;
+
+    /**
+     * Retrieve the value for the passed key
+     *
+     * @param {String} key
+     * @param {Object} options
+     * @returns {Promise<Buffer>}
+     */
+    get(key: string, options?: Object): Promise<Buffer>;
+
+    remove(key: string, options?: Object): Promise<boolean>;
+}
