+++ lib/storage/src/Adapter.js
@@ -18,9 +18,6 @@
 
 //TODO::change to interface
 class Adapter {
-    constructor(options = {}) {
-    }
-
     async open() {
         return Promise.resolve();
     }
