+++ lib/storage/src/GoogleDriveAdapter.js
@@ -17,8 +17,8 @@
  */
 
 import {Readable} from "stream"
-import Adapter from "./Adapter";
 import { drive } from "@googleapis/drive";
+import { Adapter } from "../index";
 
 /**
  * Google Drive client adapter for CloudStorage class.
@@ -45,7 +45,7 @@ export default class GoogleDriveAdapter extends Adapter {
                 folder_id = await this.createFolder(this.drive, this.folder);
             }
 
-            const file = await uploadToFolder(this.drive, val, folder_id);
+            const file = await this.uploadToFolder(this.drive, val, folder_id);
             await this.shareFile(this.drive, file.data.id);
 
             const webContentLink = await this.drive.files.get({
@@ -53,7 +53,6 @@ export default class GoogleDriveAdapter extends Adapter {
                 fields: 'webContentLink'
             }).then(response => response.data.webContentLink);
 
-            // console.log('Link to download: ', webContentLink);
             const ret = {
                 url: webContentLink,
                 storage_data: "[\"GD\",\"" + `${this.GD_VERSION}` + "\",\"" + `${file.data.id}` + "\"]"
@@ -67,7 +66,7 @@ export default class GoogleDriveAdapter extends Adapter {
     }
 
     async get(key, options = {}) {
-        return await downloadFile(this.drive, key);
+        return await this.downloadFile(this.drive, key);
     }
 
     async remove(key, options = {}) {
