+++ package.json
@@ -91,6 +91,8 @@
         "@types/chai": "^4.3.0",
         "@types/node": "^14.18.11",
         "@types/sinon": "^10.0.11",
+        "@typescript-eslint/eslint-plugin": "^5.11.0",
+        "@typescript-eslint/parser": "^5.11.0",
         "assert": "^2.0.0",
         "babel-minify": "^0.5.1",
         "babel-plugin-add-module-exports": "^1.0.4",
@@ -101,6 +103,7 @@
         "cucumber-console-formatter": "^1.0.0",
         "esdoc": "^1.1.0",
         "esdoc-standard-plugin": "^1.0.0",
+        "eslint": "^8.9.0",
         "mocha": "^6.2.3",
         "prettier": "^1.19.1",
         "pretty-quick": "^1.11.1",
