+++ tsconfig.json
@@ -39,11 +39,8 @@
     // "baseUrl": "./",                       /* Base directory to resolve non-absolute module names. */
     // "paths": {},                           /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
-    "typeRoots": [
-      "node_modules/colors",
-      "node_modules/@types"
-    ],                                        /* List of folders to include type definitions from. */
-    //"types": [],                            /* Type declaration files to be included in compilation. */
+    // "typeRoots": [],                                        /* List of folders to include type definitions from. */
+    // "types": [],                            /* Type declaration files to be included in compilation. */
     // "allowSyntheticDefaultImports": true,  /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
     "esModuleInterop": true                   /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
     // "preserveSymlinks": true,              /* Do not resolve the real path of symlinks. */
